requirements.txt
numpy

matplotlib

pandas

torch 
torchvision 

botorch

scipy

tqdm

numpy

Pillow

scikit_learn